requirements.txt
certifi==2018.4.16
chardet==3.0.4
click==6.7
idna==2.7
PyYAML==3.12
requests==2.19.1
urllib3==1.23
